+++ packages/ra-core/src/controller/field/index.ts
@@ -9,3 +9,4 @@ export * from './useReferenceArrayFieldController';
 export * from './useReferenceFieldController';
 export * from './useReferenceManyFieldController';
 export * from './useReferenceOneFieldController';
+export * from './sanitizeFieldRestProps';
