+++ packages/ra-ui-materialui/src/field/sanitizeFieldRestProps.ts
@@ -1,21 +1,6 @@
-export const sanitizeFieldRestProps: (props: any) => any = ({
-    cellClassName,
-    className,
-    emptyText,
-    fullWidth,
-    headerClassName,
-    label,
-    linkType,
-    link,
-    locale,
-    record,
-    refetch,
-    resource,
-    sortable,
-    sortBy,
-    sortByOrder,
-    source,
-    textAlign,
-    translateChoice,
-    ...props
-}) => props;
+export {
+    /**
+     * @deprecated Import from `ra-core` or `react-admin` instead
+     */
+    sanitizeFieldRestProps,
+} from 'ra-core';
