+++ packages/ra-ui-materialui/src/input/sanitizeInputRestProps.ts
@@ -1,45 +1,6 @@
-export const sanitizeInputRestProps = ({
-    afterSubmit,
-    allowNull,
-    alwaysOn,
-    beforeSubmit,
-    component,
-    data,
-    defaultValue,
-    error,
-    format,
-    formatOnBlur,
-    initializeForm,
-    input,
-    isEqual,
-    isRequired,
-    label,
-    limitChoicesToValue,
-    locale,
-    meta,
-    multiple,
-    name,
-    options,
-    optionText,
-    optionValue,
-    parse,
-    record,
-    ref,
-    refetch,
-    render,
-    resource,
-    setFilter,
-    setPagination,
-    setSort,
-    shouldUnregister,
-    source,
-    submitError,
-    subscription,
-    textAlign,
-    translate,
-    translateChoice,
-    validate,
-    validateFields,
-    value,
-    ...rest
-}: any) => rest;
+export {
+    /**
+     * @deprecated Import from `ra-core` or `react-admin` instead
+     */
+    sanitizeInputRestProps,
+} from 'ra-core';
